add static_cast<int64_t> for consistency

+++ src/mongocxx/private/numeric_casting.cpp
@@ -32,10 +32,10 @@ bool size_t_to_int64_safe(const std::size_t in, int64_t& out) {
 }
 
 bool int64_to_int32_safe(const int64_t in, int32_t& out) {
-    if (in > std::numeric_limits<int32_t>::max()) {
+    if (in > static_cast<int64_t>(std::numeric_limits<int32_t>::max())) {
         return false;
     }
-    if (in < std::numeric_limits<int32_t>::min()) {
+    if (in < static_cast<int64_t>(std::numeric_limits<int32_t>::min())) {
         return false;
     }
     out = static_cast<int32_t>(in);
